Request to update

+++ folly/synchronization/Rcu.h
@@ -308,10 +308,10 @@ class rcu_token {
   rcu_token& operator=(rcu_token&& other) = default;
 
  private:
-  explicit rcu_token(uint64_t epoch) : epoch_(epoch) {}
+  explicit rcu_token(uint8_t epoch) : epoch_(epoch) {}
 
   friend class rcu_domain<Tag>;
-  uint64_t epoch_;
+  uint8_t epoch_;
 };
 
 // Defines an RCU domain.  RCU readers within a given domain block updaters
@@ -476,8 +476,7 @@ using rcu_reader = rcu_reader_domain<RcuTag>;
 
 template <typename Tag = RcuTag>
 inline void swap(
-    rcu_reader_domain<Tag>& a,
-    rcu_reader_domain<Tag>& b) noexcept {
+    rcu_reader_domain<Tag>& a, rcu_reader_domain<Tag>& b) noexcept {
   a.swap(b);
 }
 
@@ -520,9 +519,7 @@ template <
     typename D = std::default_delete<T>,
     typename Tag = RcuTag>
 void rcu_retire(
-    T* p,
-    D d = {},
-    rcu_domain<Tag>* domain = rcu_default_domain()) {
+    T* p, D d = {}, rcu_domain<Tag>* domain = rcu_default_domain()) {
   domain->call([p, del = std::move(d)]() { del(p); });
 }
